gpu: nvgpu: Use 64-bit math for cacheline offset
Cast cacheline_start to u64 to use 64-bit maths for cacheline offset.

Coverity ID 24250

Reviewed-on: http://git-master/r/1172053

+++ drivers/gpu/nvgpu/gk20a/mm_gk20a.c
@@ -2205,7 +2205,7 @@ int gk20a_vm_map_compbits(struct vm_gk20a *vm,
 		cacheline_start = mapped_buffer->ctag_offset /
 			g->gr.comptags_per_cacheline;
 		cacheline_offset_start =
-			cacheline_start * aggregate_cacheline_sz;
+			(u64)cacheline_start * aggregate_cacheline_sz;
 
 		if (fixed_mapping) {
 			struct buffer_attrs bfr;
